ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.3",
    "fastapi>=0.120.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.10",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.2",
    "notebook>=7.4.7",
    "openai>=2.6.1",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
